Topography/XShape: store floating point offset to map center
Make ShapePoint use "float" (compatible with GL_FLOAT) and store
offsets relative to the map center instead of relative to the shape
center.  This means all XShapes live in the same coordinate system,
and they all use the same transformation matrix.  It will allow using
a single OpenGL draw command for all shapes in a file.

+++ src/Screen/OpenGL/Triangulate.hpp
@@ -49,7 +49,7 @@ PolygonToTriangles(const RasterPoint *points, unsigned num_points,
                    unsigned min_distance=1);
 unsigned
 PolygonToTriangles(const ShapePoint *points, unsigned num_points,
-                   GLushort *triangles, unsigned min_distance=1);
+                   GLushort *triangles, float min_distance=1);
 
 /**
  * Pack triangle indices into a triangle strip.
